airbnb_clone/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
importlib-metadata==1.7.0
Pillow==8.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.1.0
